+++ commands/show.go
@@ -146,7 +146,7 @@ func showResourceValuesOnlyFor(resource *graph.Resource, propKeys []string) {
 
 func showResource(resource *graph.Resource, gph *graph.Graph) {
 	displayer, err := console.BuildOptions(
-		console.WithHeaders(console.DefaultsColumnDefinitions[resource.Type()]),
+		console.WithColumnDefinitions(console.DefaultsColumnDefinitions[resource.Type()]),
 		console.WithFormat(listingFormat),
 		console.WithMaxWidth(console.GetTerminalWidth()),
 	).SetSource(resource).Build()