+++ console/humanize_test.go
@@ -22,22 +22,21 @@ import (
 )
 
 func TestHumanizeTime(t *testing.T) {
-	now := time.Now().UTC()
 	tcases := []struct {
 		stamp  time.Time
 		expect string
 	}{
-		{stamp: now, expect: "now"},
-		{stamp: now.Add(-5 * time.Second), expect: "5 secs"},
-		{stamp: now.Add(-1 * time.Minute), expect: "60 secs"},
-		{stamp: now.Add(-3 * time.Minute), expect: "3 mins"},
-		{stamp: now.Add(-90 * time.Minute), expect: "90 mins"},
-		{stamp: now.Add(-3 * time.Hour), expect: "3 hours"},
-		{stamp: now.Add(-24 * time.Hour), expect: "24 hours"},
-		{stamp: now.Add(-3 * 24 * time.Hour), expect: "3 days"},
-		{stamp: now.Add(-3 * 7 * 24 * time.Hour), expect: "3 weeks"},
-		{stamp: now.Add(-3 * 30 * 24 * time.Hour), expect: "3 months"},
-		{stamp: now.Add(-3 * 365 * 24 * time.Hour), expect: "3 years"},
+		{stamp: globalNow, expect: "now"},
+		{stamp: globalNow.Add(-5 * time.Second), expect: "5 secs"},
+		{stamp: globalNow.Add(-1 * time.Minute), expect: "60 secs"},
+		{stamp: globalNow.Add(-3 * time.Minute), expect: "3 mins"},
+		{stamp: globalNow.Add(-90 * time.Minute), expect: "90 mins"},
+		{stamp: globalNow.Add(-3 * time.Hour), expect: "3 hours"},
+		{stamp: globalNow.Add(-24 * time.Hour), expect: "24 hours"},
+		{stamp: globalNow.Add(-3 * 24 * time.Hour), expect: "3 days"},
+		{stamp: globalNow.Add(-3 * 7 * 24 * time.Hour), expect: "3 weeks"},
+		{stamp: globalNow.Add(-3 * 30 * 24 * time.Hour), expect: "3 months"},
+		{stamp: globalNow.Add(-3 * 365 * 24 * time.Hour), expect: "3 years"},
 	}
 
 	for _, tcase := range tcases {